+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - **Fix** : Fixed date syntax on split payment screens
 - **Fix** : Fixed enabling option for status field on split payment screens
 - **Fix** : Fixed browser issue about `<iframe>` HTML tag
+- **Fix** : Fixed invoice update when HiPay notification is processed before invoice creation
 
 ## 1.16.0